Fixes 'Connection refused' issue when running the example docker file.
The example docker-compose file is failing to publish to elasticmq. The log output
shows the message 'Connection refused'. This is because the Caml aws component will
first list the queues to get their urls. Elasticmq has not been configured with a
different host name so it will return "http://localhost:9324/queue/queue1" instead of
"http://sqs:9324/queue/queue1".